FastAPIProject01/requirements.txt
scikit-learn
matplotlib
xgboost
fastAPI[all]
streamlit
plotly_express
